refactor: Simulation 부분 ViewModel에서 수행하도록 변경
