+++ app/src/main/java/com/ellie/billiardsgame/ui/MainViewModel.kt
@@ -1,9 +1,11 @@
 package com.ellie.billiardsgame.ui
 
 import androidx.lifecycle.ViewModel
+import com.ellie.billiardsgame.FRAME_DURATION_MS
 import com.ellie.billiardsgame.data.Ball
 import com.ellie.billiardsgame.data.Boundary
 import com.ellie.billiardsgame.data.Point
+import java.util.concurrent.Executors
 
 class MainViewModel : ViewModel() {
     var ballDiameter = 0f
@@ -12,6 +14,9 @@ class MainViewModel : ViewModel() {
     private val redBall1 = Ball()
     private val redBall2 = Ball()
     private var boundary = Boundary()
+
+    private val executor = Executors.newFixedThreadPool(3)
+    private var isSimulating = false
 
     fun setWhiteBallPosition(x: Float, y: Float) {
         whiteBall.move(x, y)
@@ -29,7 +34,25 @@ class MainViewModel : ViewModel() {
         boundary = Boundary(Point(left.toFloat(), top.toFloat()), Point(right.toFloat(), bottom.toFloat()))
     }
 
-    fun whiteBallUpdate() {
+    fun whiteBallUpdate(x: Float, y: Float) {
+        val newX = boundary.adjustX(x, ballDiameter)
+        val newY = boundary.adjustY(y, ballDiameter)
+
+        whiteBall.move(newX, newY)
+    }
+
+    fun startSimulation() {
+        isSimulating = true
+
+        executor.submit {
+            while(isSimulating) {
+                whiteBallUpdate()
+                Thread.sleep(FRAME_DURATION_MS)
+            }
+        }
+    }
+
+    private fun whiteBallUpdate() {
         whiteBall.decreaseVelocityX()
         whiteBall.decreaseVelocityY()
 
@@ -39,11 +62,8 @@ class MainViewModel : ViewModel() {
         whiteBall.move(newX, newY)
     }
 
-    fun whiteBallUpdate(x: Float, y: Float) {
-        val newX = boundary.adjustX(x, ballDiameter)
-        val newY = boundary.adjustY(y, ballDiameter)
-
-        whiteBall.move(newX, newY)
+    fun stopSimulation() {
+        isSimulating = false
     }
 
 }
